feat: use tab to select two type of sliders
